[core]Modularize concat.
