+++ tfjs-core/src/ops/concat_split.ts
@@ -17,13 +17,11 @@
 
 import {ENGINE} from '../engine';
 import {Tensor, Tensor1D, Tensor2D, Tensor3D, Tensor4D} from '../tensor';
-import {convertToTensor, convertToTensorArray} from '../tensor_util_env';
+import {convertToTensor} from '../tensor_util_env';
 import {TensorLike} from '../types';
-import {assert, sizeFromShape} from '../util';
+import {assert,} from '../util';
 import {parseAxisParam} from '../util';
-import {assertParamsConsistent, computeOutShape} from './concat_util';
 import {op} from './operation';
-import {tensor} from './tensor_ops';
 
 /**
  * Concatenates a list of`tf.Tensor1D`s along an axis. See `concat` for details.
@@ -122,82 +120,6 @@ function concat4d_(
 }
 
 /**
- * Concatenates a list of `tf.Tensor`s along a given axis.
- *
- * The tensors ranks and types must match, and their sizes must match in all
- * dimensions except `axis`.
- *
- * Also available are stricter rank-specific methods that assert that
- * `tensors` are of the given rank:
- *   - `tf.concat1d`
- *   - `tf.concat2d`
- *   - `tf.concat3d`
- *   - `tf.concat4d`
- *
- * Except `tf.concat1d` (which does not have axis param), all methods have
- * same signature as this method.
- *
- * ```js
- * const a = tf.tensor1d([1, 2]);
- * const b = tf.tensor1d([3, 4]);
- * a.concat(b).print();  // or a.concat(b)
- * ```
- *
- * ```js
- * const a = tf.tensor1d([1, 2]);
- * const b = tf.tensor1d([3, 4]);
- * const c = tf.tensor1d([5, 6]);
- * tf.concat([a, b, c]).print();
- * ```
- *
- * ```js
- * const a = tf.tensor2d([[1, 2], [10, 20]]);
- * const b = tf.tensor2d([[3, 4], [30, 40]]);
- * const axis = 1;
- * tf.concat([a, b], axis).print();
- * ```
- * @param tensors A list of tensors to concatenate.
- * @param axis The axis to concate along. Defaults to 0 (the first dim).
- */
-/** @doc {heading: 'Tensors', subheading: 'Slicing and Joining'} */
-function concat_<T extends Tensor>(tensors: Array<T|TensorLike>, axis = 0): T {
-  assert(tensors.length >= 1, () => 'Pass at least one tensor to concat');
-  let $tensors = convertToTensorArray(tensors, 'tensors', 'concat');
-  if ($tensors[0].dtype === 'complex64') {
-    $tensors.forEach(tensor => {
-      if (tensor.dtype !== 'complex64') {
-        throw new Error(`Cannot concatenate complex64 tensors with a tensor
-          with dtype ${tensor.dtype}. `);
-      }
-    });
-  }
-
-  axis = parseAxisParam(axis, $tensors[0].shape)[0];
-  const outShape = computeOutShape($tensors.map(t => t.shape), axis);
-  if (sizeFromShape(outShape) === 0) {
-    return tensor([], outShape) as T;
-  }
-  // Keep only non-empty tensors (ignore tensors with 0 in their shape).
-  $tensors = $tensors.filter(t => t.size > 0);
-  if ($tensors.length === 1) {
-    return $tensors[0];
-  }
-
-  const shapes = $tensors.map(t => t.shape);
-  assertParamsConsistent(shapes, axis);
-  const der = (dy: T) => {
-    const sizeSplits = shapes.map(s => s[axis]);
-    const derTensors = split(dy, sizeSplits, axis);
-    return derTensors.map(t => () => t) as {};
-  };
-  const inputs = $tensors as {};
-  const attr = {axis};
-  return ENGINE.runKernelFunc(
-      backend => backend.concat($tensors, axis) as T, inputs, der, 'Concat',
-      attr);
-}
-
-/**
  * Splits a `tf.Tensor` into sub tensors.
  *
  * If `numOrSizeSplits` is a number, splits `x` along dimension `axis`
@@ -253,7 +175,6 @@ function split_<T extends Tensor>(
       backend => backend.split($x, splitSizes, axis), {$x}, der);
 }
 
-export const concat = op({concat_});
 export const concat1d = op({concat1d_});
 export const concat2d = op({concat2d_});
 export const concat3d = op({concat3d_});
